requirements.txt
Django==2.0.5
django-model-utils==3.1.2
django-pipeline==1.6.14
